src/pyproject.toml
[project]
name = "src"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.6.0",
    "art>=6.5",
    "black>=25.1.0",
    "colorama>=0.4.6",
    "google-genai>=1.13.0",
    "numpy>=2.2.5",
    "openai>=1.76.2",
    "pillow>=11.2.1",
    "pyfiglet>=1.0.2",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.1.0",
    "rich>=14.0.0",
    "scikit-image>=0.25.2",
    "scikit-learn>=1.6.1",
    "torch>=2.7.0",
    "transformers>=4.51.3",
    "uvicorn>=0.34.2",
    "yaspin>=3.1.0",
]
